Corrections after feedback
Titles and axis label were added to the chart, unfortunately the chart
library does not support labels in the used chart types.
Added descriptions and explanations in the notebook file.
Refactor of the data web end point to be more idiomatic
Refactor of the spider bots to the correct way to get arguments from the
command line.